+++ src/components/modules/mainBit/styles.js
