+++ src/redux/modules/updateWord/reducer.js
@@ -1,5 +1,4 @@
-
-import wordsArray from '../../../components/greetings';
+import wordsArray from '../../../components/modules/greetings';
 
 export const currentWordReducer = (state = wordsArray[2], action) => {
   switch (action.type) {
